New ID picking system, default wait time changed, andsyntax improvement (v2.0)
FERUN will no longer have a goal; instead it will be infinite.
First, it will pick an era (like 2004-2007), then obtain an ID from that era and tweet the username that matches an ID that matches the picked era.
If the '2016-2020' era is picked, it'll generate an ID between the first registered 2016 ID and the last ID registered (that can be obtained via RecentRobloxUsers.py).
If RecentRobloxUsers.py is not being run, it could use the

Due to this, maximum count, minimum and maximum ID has been removed from the config.cfg file.

Default wait time has been changed to 5 minutes.